Added authentication for users.

+++ server/routes/timeline.router.js
@@ -6,6 +6,7 @@ const router = express.Router();
  * GET route template
  */
 router.get('/', (req, res) => {
+  if (req.isAuthenticated()) {
   queryText = `SELECT * FROM "timeline"
   WHERE "user_id" = $1 ORDER BY "date_created";`;
   pool.query(queryText, [req.user.id]).then((result) => {
@@ -14,13 +15,16 @@ router.get('/', (req, res) => {
   }).catch((err) => {
     console.log('GET timeline error', err)
     res.sendStatus(500);
-  })
+  })} else {
+    res.sendStatus(401);
+  }
 });
 
 /**
  * POST route template
  */
 router.post('/', (req, res) => {
+  if(req.isAuthenticated()) {
     queryText = `INSERT INTO "timeline" ("title", "date_created", "user_id")
     VALUES (
       $1, CURRENT_TIMESTAMP, $2
@@ -31,11 +35,14 @@ router.post('/', (req, res) => {
     }).catch((e) => {
       console.error(e);
       res.sendStatus(500);
-    })
+    })} else {
+      res.sendStatus(401);
+    }
   });
 
 // PUT Route
   router.put('/:id', (req, res) => {
+    if(req.isAuthenticated()) {
     let queryText = `UPDATE "timeline"
     SET "visible" = NOT "visible"
     WHERE "id" = $1 AND "user_id" = $2;`;
@@ -46,12 +53,14 @@ router.post('/', (req, res) => {
     }).catch((e) => {
         console.error(e);
         res.sendStatus(500);
-    })
+    })} else {
+      res.sendStatus(401);
+    }
 })
 
 // DELETE Route
 router.delete('/:id', (req, res) => {
-  //! THIS MUST DELETE ALL POSTS WITH THE TIMELINE ID FOR IT TO WORK
+  if (req.isAuthenticated()) {
     let queryText = `DELETE FROM "timeline"
       WHERE "id" = $1 AND "user_id" = $2;`;
     pool.query(queryText, [req.params.id, req.user.id])
@@ -60,7 +69,9 @@ router.delete('/:id', (req, res) => {
     }).catch((err) => {
         console.error('deletion error', err)
         res.sendStatus(500)
-    })
+    })} else {
+      res.sendStatus(401);
+    }
 })
